% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectProtterAPI.R
\name{connectProtterAPI}
\alias{connectProtterAPI}
\title{Title}
\usage{
connectProtterAPI(
  evidence = NULL,
  SelectedExperiment = NULL,
  SelectedProtein = NULL,
  combineExperiments = FALSE,
  plot_palette = FALSE,
  peptideCutter = TRUE
)
}
\arguments{
\item{SelectedProtein}{}
}
\value{

}
\description{
Title
}
