% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPTMs.R
\name{createPTMs}
\alias{createPTMs}
\title{Title}
\usage{
createPTMs(
  evidence,
  peptideType = c("Unmodified", "Modified", "Both"),
  selectedProtein,
  selectedExperiment,
  experimentDesign = NULL,
  selectedCondition = NULL,
  plotLegend = TRUE
)
}
\arguments{
\item{plotLegend}{}
}
\value{

}
\description{
Title
}
