% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinCoverage.R
\name{proteinCoverage}
\alias{proteinCoverage}
\title{Title}
\usage{
proteinCoverage(
  proteinId,
  dfPeptidesColors,
  comparison = c("individualExperiments", "combineExperiments", "conditions")
)
}
\arguments{
\item{evidence}{}
}
\value{

}
\description{
Title
}
