% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonPTMs.R
\name{comparisonPTMs}
\alias{comparisonPTMs}
\title{Title}
\usage{
comparisonPTMs(
  evidence,
  peptideType = c("Unmodified", "Modified", "Both"),
  selectedProtein,
  selectedExperiment,
  experimentDesign = NULL,
  selectedConditions = NULL
)
}
\arguments{
\item{selectedCondition}{}
}
\value{

}
\description{
Title
}
