% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectProtterAPI.R
\name{connectProtterAPI}
\alias{connectProtterAPI}
\title{Title}
\usage{
connectProtterAPI(
  dfPeptidesColors = NULL,
  SelectedProtein = NULL,
  proteaseSelected = "Tryps",
  openInProtter = FALSE
)
}
\arguments{
\item{proteaseSelected}{}
}
\value{

}
\description{
Title
}
