% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectProtterAPI.R
\name{connectProtterAPI}
\alias{connectProtterAPI}
\title{Title}
\usage{
connectProtterAPI(
  evidence,
  SelectedExperiment,
  SelectedProtein,
  combineExperiments = FALSE
)
}
\arguments{
\item{combineExperiments}{}
}
\value{

}
\description{
Title
}
