% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLegend.R
\name{createLegend}
\alias{createLegend}
\title{Title}
\usage{
createLegend(
  evidence = NULL,
  SelectedExperiment = NULL,
  SelectedProtein = NULL,
  combineExperiments = FALSE,
  plot_legend = TRUE,
  proteaseSelected = "Tryps"
)
}
\arguments{
\item{proteaseSelected}{}
}
\value{

}
\description{
Title
}
