% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLegend.R
\name{createLegend}
\alias{createLegend}
\title{Title}
\usage{
createLegend(
  evidence,
  experimentDesign,
  comparison = c("individualExperiments", "combineExperiments", "conditions"),
  conditionSelected = NULL,
  selectedExperiment = NULL,
  selectedProtein = NULL,
  plot_legend = TRUE
)
}
\arguments{
\item{conditionSelected}{}
}
\value{

}
\description{
Title
}
