% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLegend.R
\name{createLegend}
\alias{createLegend}
\title{Title}
\usage{
createLegend(
  evidence,
  experimentDesign,
  compareConditions = FALSE,
  conditionSelected = NULL,
  SelectedExperiment = NULL,
  selectedProtein = NULL,
  combineExperiments = FALSE,
  plot_legend = TRUE
)
}
\arguments{
\item{plot_legend}{}
}
\value{

}
\description{
Title
}
